requirements.txt
#translator/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
openai
google-genai
